Add --mode flag to allow running offline or not sending package list
Using --mode share-nothing, the whole data.json file is
retrieved from the server, without sharing the list of
packages from node_modules to the server.

Using --mode offline, the bundled data.json is used.

Finally, --mode online is the default, and sends the list
of packages to the server and retrieves the filtered data.